projects/davi-marques/requirements.txt
telebot
Flask
Pillow